+++ scripts/ci_sudo/rsnapshot_backup_sync.sh
@@ -37,11 +37,21 @@ exec 2>&1
 set -x
 set -o pipefail
 if [[ "${RSNAPSHOT_BACKUP_TYPE}" == "SSH" ]]; then
-	ssh -o BatchMode=yes -o StrictHostKeyChecking=no ${SSH_JUMP} -p ${SSH_PORT} ${SSH_HOST} \
-		"bash -c 'exec > >(tee /opt/sysadmws/rsnapshot_backup/rsnapshot_backup.log); exec 2>&1; /opt/sysadmws/rsnapshot_backup/rsnapshot_backup.sh sync'" | ccze -A || GRAND_EXIT=1
+	if salt ${TARGET} pillar.get rsnapshot_backup:python | grep -q -e True; then
+		ssh -o BatchMode=yes -o StrictHostKeyChecking=no ${SSH_JUMP} -p ${SSH_PORT} ${SSH_HOST} \
+			"bash -c 'exec > >(tee /opt/sysadmws/rsnapshot_backup/rsnapshot_backup.log); exec 2>&1; /opt/sysadmws/rsnapshot_backup/rsnapshot_backup.py --sync'" | ccze -A || GRAND_EXIT=1
+	else
+		ssh -o BatchMode=yes -o StrictHostKeyChecking=no ${SSH_JUMP} -p ${SSH_PORT} ${SSH_HOST} \
+			"bash -c 'exec > >(tee /opt/sysadmws/rsnapshot_backup/rsnapshot_backup.log); exec 2>&1; /opt/sysadmws/rsnapshot_backup/rsnapshot_backup.sh sync'" | ccze -A || GRAND_EXIT=1
+	fi
 elif [[ "${RSNAPSHOT_BACKUP_TYPE}" == "SALT" ]]; then
-	salt --force-color -t ${SALT_TIMEOUT} ${TARGET} cmd.run \
-		"bash -c 'exec > >(tee /opt/sysadmws/rsnapshot_backup/rsnapshot_backup.log); exec 2>&1; /opt/sysadmws/rsnapshot_backup/rsnapshot_backup.sh sync'" | ccze -A || GRAND_EXIT=1
+	if salt ${TARGET} pillar.get rsnapshot_backup:python | grep -q -e True; then
+		salt --force-color -t ${SALT_TIMEOUT} ${TARGET} cmd.run \
+			"bash -c 'exec > >(tee /opt/sysadmws/rsnapshot_backup/rsnapshot_backup.log); exec 2>&1; /opt/sysadmws/rsnapshot_backup/rsnapshot_backup.py --sync'" | ccze -A || GRAND_EXIT=1
+	else
+		salt --force-color -t ${SALT_TIMEOUT} ${TARGET} cmd.run \
+			"bash -c 'exec > >(tee /opt/sysadmws/rsnapshot_backup/rsnapshot_backup.log); exec 2>&1; /opt/sysadmws/rsnapshot_backup/rsnapshot_backup.sh sync'" | ccze -A || GRAND_EXIT=1
+	fi
 else
 	echo ERROR: unknown RSNAPSHOT_BACKUP_TYPE
 	exit 1
